# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Steel Fabrication MRP API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /jobs/{jobId}/documents:
    get:
      operationId: listJobDocuments
      tags: [documents]
      summary: List documents attached to a job
      parameters:
        - $ref: "#/components/parameters/JobId"
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: uploadJobDocument
      tags: [documents]
      summary: Upload a document to a job
      description: >
        Multipart upload. Fields: `file` (binary) and `category`
        (drawing, mtr, photo, nc_data, spreadsheet, other).
        Allowed file types: PDF, DWG, DXF, NC-1/DSTV (.nc1, .nc), JPG, PNG, XLSX, CSV.
      parameters:
        - $ref: "#/components/parameters/JobId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadForm"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input (disallowed file type or too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /estimates/{estimateId}/documents:
    get:
      operationId: listEstimateDocuments
      tags: [documents]
      summary: List documents attached to an estimate
      parameters:
        - $ref: "#/components/parameters/EstimateId"
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: uploadEstimateDocument
      tags: [documents]
      summary: Upload a document to an estimate
      description: >
        Multipart upload. Fields: `file` (binary) and `category`
        (drawing, mtr, photo, nc_data, spreadsheet, other).
        Allowed file types: PDF, DWG, DXF, NC-1/DSTV (.nc1, .nc), JPG, PNG, XLSX, CSV.
      parameters:
        - $ref: "#/components/parameters/EstimateId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadForm"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input (disallowed file type or too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /documents/{documentId}:
    delete:
      operationId: deleteDocument
      tags: [documents]
      summary: Delete a document
      description: Removes the document record and its stored file.
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /documents/{documentId}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Download a document file
      parameters:
        - $ref: "#/components/parameters/DocumentId"
      responses:
        "200":
          description: File contents
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bom/parts/{partId}/documents:
    get:
      operationId: listPartDocuments
      tags: [documents]
      summary: List documents attached to a BOM part
      parameters:
        - $ref: "#/components/parameters/PartId"
      responses:
        "200":
          description: Documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    post:
      operationId: uploadPartDocument
      tags: [documents]
      summary: Upload a PDF to a BOM part
      description: >
        Multipart upload of a mill test report or similar certificate for a
        specific part line. Field: `file` (binary). PDF files only.
        The document is stored with category `mtr`.
      parameters:
        - $ref: "#/components/parameters/PartId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/PartDocumentUploadForm"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
        "400":
          description: Invalid input (not a PDF or too large)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /time-clock/out:
    post:
      operationId: clockOut
      tags: [time]
      summary: Clock out
      description: Shop floor clock out for an open time entry.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClockOutInput"
      responses:
        "200":
          description: Clocked out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimeEntry"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  parameters:
...
    JobId:
      name: jobId
      in: path
      required: true
      schema:
        type: integer
...
    PartId:
      name: partId
      in: path
      required: true
      schema:
        type: integer
...
    DocumentId:
      name: documentId
      in: path
      required: true
      schema:
        type: integer
    EstimateId:
      name: estimateId
      in: path
      required: true
      schema:
        type: integer
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    TimeEntry:
      type: object
      required:
        - id
        - employeeId
        - employeeName
        - jobId
        - jobNumber
        - jobName
        - stageId
        - stageName
        - clockIn
        - createdAt
      properties:
        id:
          type: integer
        employeeId:
          type: integer
        employeeName:
          type: string
        employeeTitle:
          type: ["string", "null"]
        jobId:
          type: integer
        jobNumber:
          type: string
        jobName:
          type: string
        stageId:
          type: integer
        stageName:
          type: string
        clockIn:
          type: string
          format: date-time
        clockOut:
          type: ["string", "null"]
          format: date-time
        durationMinutes:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    ClockOutInput:
      type: object
      required: [entryId]
      properties:
        entryId:
          type: integer
...
    DocumentCategory:
      type: string
      enum: [drawing, mtr, photo, nc_data, spreadsheet, other]
    DocumentUploadForm:
      type: object
      required: [file, category]
      properties:
        file:
          type: string
          format: binary
        category:
          $ref: "#/components/schemas/DocumentCategory"
    PartDocumentUploadForm:
      type: object
      required: [file]
      properties:
        file:
          type: string
          format: binary
    Document:
      type: object
      required:
        - id
        - jobId
        - estimateId
        - partId
        - filename
        - category
        - mimeType
        - sizeBytes
        - uploadedAt
      properties:
        id:
          type: integer
        jobId:
          type: ["integer", "null"]
        estimateId:
          type: ["integer", "null"]
        partId:
          type: ["integer", "null"]
        filename:
          type: string
        category:
          $ref: "#/components/schemas/DocumentCategory"
        mimeType:
          type: string
        sizeBytes:
          type: integer
        uploadedAt:
          type: string
          format: date-time
...